Fix data type of osync_group_find_member() parameter member id
refs #974



git-svn-id: http://svn.opensync.org/trunk@5454 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e

+++ opensync/group/opensync_group.c
@@ -788,7 +788,7 @@ void osync_group_remove_member(OSyncGroup *group, OSyncMember *member)
 	osync_member_unref(member);
 }
 
-OSyncMember *osync_group_find_member(OSyncGroup *group, int id)
+OSyncMember *osync_group_find_member(OSyncGroup *group, long long int id)
 {
 	OSyncList *m = NULL;
 	for (m = group->members; m; m = m->next) {
